Refactored OrderedRDD2 to take a RegionValueRDD
